Add a text index for tweets
I want to be able to efficiently search for tweets and a text index
seems like a good way to do this.

To make this explicit, I created some JavaScript scripts that can be run
in the mongo shell to add and remove the index.  The usage of the
migrations is documented in the README.

Addresses #8