Add MongoDB 3.4+ as an assumption
Trying to deploy changes for #8 on production, I found the search
didn't work.  It looks like text search isn't supported (or supported)
in the same way on earlier versions of mongo.  I had to install the
mongodb .deb from the external repo following the instructions at
https://docs.mongodb.com/manual/tutorial/install-mongodb-on-ubuntu/
because the official packages for Ubuntu 16.04 include the
no-longer-supported MongoDB 2.6.10.